projects/1_init_and_consume/pyproject.toml
[project]
name = "example-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]